Crate books1

This documentation was generated from books crate version 0.0.1. The original source code can be found on github.

Features

Handle the following Resources with ease from the central hub ...

bookshelves (get, list and volumes.list)

cloudloading (add book, delete book and update book)

dictionary (list offline metadata)

layers (annotation data.get, annotation data.list, get, list, volume annotations.get and volume annotations.list)

myconfig (get user settings, release download access, request access, sync volume licenses and update user settings)

mylibrary (annotations.delete, annotations.get, annotations.insert, annotations.list, annotations.summary, annotations.update, bookshelves.add volume, bookshelves.clear volumes, bookshelves.get, bookshelves.list, bookshelves.move volume, bookshelves.remove volume, bookshelves.volumes.list, readingpositions.get and readingpositions.set position)

onboarding (list categories and list category volumes)

promooffer (accept, dismiss and get)

volumes (associated.list, get, list, mybooks.list, recommended.list, recommended.rate and useruploaded.list)

Everything else about the books v1 API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.volumes().useruploaded_list(...).doit()
let r = hub.myconfig().sync_volume_licenses(...).doit()
let r = hub.volumes().list(...).doit()
let r = hub.volumes().associated_list(...).doit()
let r = hub.bookshelves().volumes_list(...).doit()
let r = hub.volumes().recommended_list(...).doit()
let r = hub.mylibrary().bookshelves_volumes_list(...).doit()
let r = hub.volumes().mybooks_list(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
books1 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate books1;
use books1::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use books1::Books;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Books::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.volumes().list("q")
             .start_index(11)
             .source("dolores")
             .show_preorders(false)
             .projection("sadipscing")
             .print_type("aliquyam")
             .partner("ea")
             .order_by("no")
             .max_results(80)
             .library_restrict("justo")
             .lang_restrict("et")
             .filter("et")
             .download("diam")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

Annotation
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

annotations.insert mylibrary (request|response)

annotations.get mylibrary (response)

annotations.update mylibrary (request|response)

AnnotationClientVersionRanges
Selection ranges sent from the client.

This type is not used in any activity, and only used as part of another schema.

AnnotationCurrentVersionRanges
Selection ranges for the most recent content version.

This type is not used in any activity, and only used as part of another schema.

AnnotationLayerSummary
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Annotationdata
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

annotation data.get layers (response)

Annotations
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

annotations.list mylibrary (response)

AnnotationsSummary
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

annotations.summary mylibrary (response)

AnnotationsSummaryLayers
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Annotationsdata
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

annotation data.list layers (response)

Books
Central instance to access all Books related resource activities

BooksAnnotationsRange
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

BooksCloudloadingResource
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update book cloudloading (request|response)

add book cloudloading (response)

BooksVolumesRecommendedRateResponse
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

recommended.rate volumes (response)

Bookshelf
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get bookshelves (response)

bookshelves.get mylibrary (response)

BookshelveGetMethodBuilder
Retrieves metadata for a specific bookshelf for the specified user.

BookshelveListMethodBuilder
Retrieves a list of public bookshelves for the specified user.

BookshelveMethodsBuilder
A builder providing access to all methods supported on bookshelve resources. It is not used directly, but through the Books hub.

BookshelveVolumeListMethodBuilder
Retrieves volumes in a specific bookshelf for the specified user.

Bookshelves
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

bookshelves.list mylibrary (response)

list bookshelves (response)

Category
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list categories onboarding (response)

CategoryItems
A list of onboarding categories.

This type is not used in any activity, and only used as part of another schema.

CloudloadingAddBookMethodBuilder

CloudloadingDeleteBookMethodBuilder
Remove the book and its contents

CloudloadingMethodsBuilder
A builder providing access to all methods supported on cloudloading resources. It is not used directly, but through the Books hub.

CloudloadingUpdateBookMethodBuilder

ConcurrentAccessRestriction
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

DictionaryListOfflineMetadataMethodBuilder
Returns a list of offline dictionary meatadata available

DictionaryMethodsBuilder
A builder providing access to all methods supported on dictionary resources. It is not used directly, but through the Books hub.

Dictlayerdata
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

DictlayerdataCommon
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

DictlayerdataDict
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

DictlayerdataDictSource
The source, url and attribution for this dictionary data.

This type is not used in any activity, and only used as part of another schema.

DictlayerdataDictWords
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

DictlayerdataDictWordsDerivatives
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

DictlayerdataDictWordsDerivativesSource
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

DictlayerdataDictWordsExamples
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

DictlayerdataDictWordsExamplesSource
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

DictlayerdataDictWordsSenses
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

DictlayerdataDictWordsSensesConjugations
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

DictlayerdataDictWordsSensesDefinitions
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

DictlayerdataDictWordsSensesDefinitionsExamples
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

DictlayerdataDictWordsSensesDefinitionsExamplesSource
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

DictlayerdataDictWordsSensesSource
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

DictlayerdataDictWordsSensesSynonyms
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

DictlayerdataDictWordsSensesSynonymsSource
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

DictlayerdataDictWordsSource
The words with different meanings but not related words, e.g. "go" (game) and "go" (verb).

This type is not used in any activity, and only used as part of another schema.

DownloadAccessRestriction
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

DownloadAccesses
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

release download access myconfig (response)

Geolayerdata
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

GeolayerdataCommon
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

GeolayerdataGeo
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

GeolayerdataGeoBoundary
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

GeolayerdataGeoViewport
The viewport for showing this location. This is a latitude, longitude rectangle.

This type is not used in any activity, and only used as part of another schema.

GeolayerdataGeoViewportHi
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

GeolayerdataGeoViewportLo
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

LayerAnnotationDataGetMethodBuilder
Gets the annotation data.

LayerAnnotationDataListMethodBuilder
Gets the annotation data for a volume and layer.

LayerGetMethodBuilder
Gets the layer summary for a volume.

LayerListMethodBuilder
List the layer summaries for a volume.

LayerMethodsBuilder
A builder providing access to all methods supported on layer resources. It is not used directly, but through the Books hub.

LayerVolumeAnnotationGetMethodBuilder
Gets the volume annotation.

LayerVolumeAnnotationListMethodBuilder
Gets the volume annotations for a volume and layer.

Layersummaries
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list layers (response)

Layersummary
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get layers (response)

Metadata
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list offline metadata dictionary (response)

MetadataItems
A list of offline dictionary metadata.

This type is not used in any activity, and only used as part of another schema.

MyconfigGetUserSettingMethodBuilder
Gets the current settings for the user.

MyconfigMethodsBuilder
A builder providing access to all methods supported on myconfig resources. It is not used directly, but through the Books hub.

MyconfigReleaseDownloadAccesMethodBuilder
Release downloaded content access restriction.

MyconfigRequestAccesMethodBuilder
Request concurrent and download access restrictions.

MyconfigSyncVolumeLicenseMethodBuilder
Request downloaded content access for specified volumes on the My eBooks shelf.

MyconfigUpdateUserSettingMethodBuilder
Sets the settings for the user. Unspecified sub-objects will retain the existing value.

MylibraryAnnotationDeleteMethodBuilder
Deletes an annotation.

MylibraryAnnotationGetMethodBuilder
Gets an annotation by its ID.

MylibraryAnnotationInsertMethodBuilder
Inserts a new annotation.

MylibraryAnnotationListMethodBuilder
Retrieves a list of annotations, possibly filtered.

MylibraryAnnotationSummaryMethodBuilder
Gets the summary of specified layers.

MylibraryAnnotationUpdateMethodBuilder
Updates an existing annotation.

MylibraryBookshelveAddVolumeMethodBuilder
Adds a volume to a bookshelf.

MylibraryBookshelveClearVolumeMethodBuilder
Clears all volumes from a bookshelf.

MylibraryBookshelveGetMethodBuilder
Retrieves metadata for a specific bookshelf belonging to the authenticated user.

MylibraryBookshelveListMethodBuilder
Retrieves a list of bookshelves belonging to the authenticated user.

MylibraryBookshelveMoveVolumeMethodBuilder
Moves a volume within a bookshelf.

MylibraryBookshelveRemoveVolumeMethodBuilder
Removes a volume from a bookshelf.

MylibraryBookshelveVolumeListMethodBuilder
Gets volume information for volumes on a bookshelf.

MylibraryMethodsBuilder
A builder providing access to all methods supported on mylibrary resources. It is not used directly, but through the Books hub.

MylibraryReadingpositionGetMethodBuilder
Retrieves my reading position information for a volume.

MylibraryReadingpositionSetPositionMethodBuilder
Sets my reading position information for a volume.

Offers
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get promooffer (response)

OffersItems
A list of offers.

This type is not used in any activity, and only used as part of another schema.

OffersItemsItems
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

OnboardingListCategoryMethodBuilder
List categories for onboarding experience.

OnboardingListCategoryVolumeMethodBuilder
List available volumes under categories for onboarding experience.

OnboardingMethodsBuilder
A builder providing access to all methods supported on onboarding resources. It is not used directly, but through the Books hub.

PromoofferAcceptMethodBuilder

PromoofferDismisMethodBuilder

PromoofferGetMethodBuilder
Returns a list of promo offers available to the user

PromoofferMethodsBuilder
A builder providing access to all methods supported on promooffer resources. It is not used directly, but through the Books hub.

ReadingPosition
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

readingpositions.get mylibrary (response)

RequestAccess
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

request access myconfig (response)

Review
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

ReviewAuthor
Author of this review.

This type is not used in any activity, and only used as part of another schema.

ReviewSource
Information regarding the source of this review, when the review is not from a Google Books user.

This type is not used in any activity, and only used as part of another schema.

Usersettings
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get user settings myconfig (response)

update user settings myconfig (request|response)

UsersettingsNotesExport
User settings in sub-objects, each for different purposes.

This type is not used in any activity, and only used as part of another schema.

Volume
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

useruploaded.list volumes (none)

recommended.rate volumes (none)

list volumes (none)

associated.list volumes (none)

get volumes (response)

recommended.list volumes (none)

mybooks.list volumes (none)

Volume2
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list category volumes onboarding (response)

VolumeAccessInfo
Any information about a volume related to reading or obtaining that volume text. This information can depend on country (books may be public domain in one country but not in another, e.g.).

This type is not used in any activity, and only used as part of another schema.

VolumeAccessInfoEpub
Information about epub content. (In LITE projection.)

This type is not used in any activity, and only used as part of another schema.

VolumeAccessInfoPdf
Information about pdf content. (In LITE projection.)

This type is not used in any activity, and only used as part of another schema.

VolumeAssociatedListMethodBuilder
Return a list of associated books.

VolumeGetMethodBuilder
Gets volume information for a single volume.

VolumeLayerInfo
What layers exist in this volume and high level information about them.

This type is not used in any activity, and only used as part of another schema.

VolumeLayerInfoLayers
A layer should appear here if and only if the layer exists for this book.

This type is not used in any activity, and only used as part of another schema.

VolumeListMethodBuilder
Performs a book search.

VolumeMethodsBuilder
A builder providing access to all methods supported on volume resources. It is not used directly, but through the Books hub.

VolumeMybookListMethodBuilder
Return a list of books in My Library.

VolumeRecommendedInfo
Recommendation related information for this volume.

This type is not used in any activity, and only used as part of another schema.

VolumeRecommendedListMethodBuilder
Return a list of recommended books for the current user.

VolumeRecommendedRateMethodBuilder
Rate a recommended book for the current user.

VolumeSaleInfo
Any information about a volume related to the eBookstore and/or purchaseability. This information can depend on the country where the request originates from (i.e. books may not be for sale in certain countries).

This type is not used in any activity, and only used as part of another schema.

VolumeSaleInfoListPrice
Suggested retail price. (In LITE projection.)

This type is not used in any activity, and only used as part of another schema.

VolumeSaleInfoOffers
Offers available for this volume (sales and rentals).

This type is not used in any activity, and only used as part of another schema.

VolumeSaleInfoOffersListPrice
Offer list (=undiscounted) price in Micros.

This type is not used in any activity, and only used as part of another schema.

VolumeSaleInfoOffersRentalDuration
The rental duration (for rental offers only).

This type is not used in any activity, and only used as part of another schema.

VolumeSaleInfoOffersRetailPrice
Offer retail (=discounted) price in Micros

This type is not used in any activity, and only used as part of another schema.

VolumeSaleInfoRetailPrice
The actual selling price of the book. This is the same as the suggested retail or list price unless there are offers or discounts on this volume. (In LITE projection.)

This type is not used in any activity, and only used as part of another schema.

VolumeSearchInfo
Search result information related to this volume.

This type is not used in any activity, and only used as part of another schema.

VolumeUserInfo
User specific information related to this volume. (e.g. page this user last read or whether they purchased this book)

This type is not used in any activity, and only used as part of another schema.

VolumeUserInfoCopy
Copy/Paste accounting information.

This type is not used in any activity, and only used as part of another schema.

VolumeUserInfoRentalPeriod
Period during this book is/was a valid rental.

This type is not used in any activity, and only used as part of another schema.

VolumeUserInfoUserUploadedVolumeInfo
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

VolumeUseruploadedListMethodBuilder
Return a list of books uploaded by the current user.

VolumeVolumeInfo
General volume information.

This type is not used in any activity, and only used as part of another schema.

VolumeVolumeInfoDimensions
Physical dimensions of this volume.

This type is not used in any activity, and only used as part of another schema.

VolumeVolumeInfoImageLinks
A list of image links for all the sizes that are available. (In LITE projection.)

This type is not used in any activity, and only used as part of another schema.

VolumeVolumeInfoIndustryIdentifiers
Industry standard identifiers for this volume.

This type is not used in any activity, and only used as part of another schema.

Volumeannotation
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

volume annotations.get layers (response)

VolumeannotationContentRanges
The content ranges to identify the selected text.

This type is not used in any activity, and only used as part of another schema.

Volumeannotations
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

volume annotations.list layers (response)

Volumes
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

useruploaded.list volumes (response)

sync volume licenses myconfig (response)

list volumes (response)

associated.list volumes (response)

volumes.list bookshelves (response)

recommended.list volumes (response)

bookshelves.volumes.list mylibrary (response)

mybooks.list volumes (response)

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.